+++ src/interface/common.ts
@@ -1,0 +1,2 @@
+export type TextFormatter<T> = (payload: T) => string;
+export type Text<T> = string | TextFormatter<T>;